+++ hikari_orchestrator/__main__.py
@@ -30,14 +30,22 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 from __future__ import annotations
 
+import importlib
 import io
 import logging
+import sys
 
 import click
 import dotenv
 import hikari
 
 from . import _service  # pyright: ignore[reportPrivateUsage]
+
+_ENV_PREFIX = "ORCHESTRATOR_"
+
+
+def _env_name(name: str) -> str:
+    return f"{_ENV_PREFIX}{name}"
 
 
 def _cast_intents(value: str, /) -> hikari.Intents:
@@ -61,47 +69,139 @@ def _cast_intents(value: str, /) -> hikari.Intents:
 
 
 _HELP = """
+Command line entry points for Hikari Orchestrator: a tool for managing Hikari shard clusters."""
+
+
+@click.group(help=_HELP, invoke_without_command=True)
+def _cli_entry() -> None:
+    ...
+
+
+_RUN_HELP = """
+Run a standalone multi-process Hikari bot on the current system.
+
+Both `--entrypoint` and `--token` must be passed (either directly or as env variables).
+"""
+
+
+@_cli_entry.command(name="run", help=_RUN_HELP)
+@click.option(
+    "--entrypoint",
+    "-ep",
+    envvar=_env_name("ENTRYPOINT"),
+    help=(
+        "Path to the function which will be called with each initialised bot object. "
+        "This must be in the format '{module_path}:{function_name}'."
+    ),
+    required=True,
+)
+@click.option("--token", envvar="DISCORD_TOKEN", help="Discord token for the bot to orchestrate.", required=True)
+@click.option(
+    "--intents",
+    default=hikari.Intents.ALL_UNPRIVILEGED,
+    envvar=_env_name("INTENTS"),
+    show_default=True,
+    help="Gateway intents the bot should use.",
+    type=_cast_intents,
+)
+@click.option(
+    "--shard-count",
+    default=None,
+    envvar=_env_name("SHARD_COUNT"),
+    help="The amount of shards to run for the bot. Defaults to Discord's recommended amount.",
+    type=int,
+)
+@click.option("--log-level", default="INFO", envvar="LOG_LEVEL", help="A Python logging level name.", show_default=True)
+@click.option(
+    "--process-count",
+    default=_service.DEFAULT_SUBPROCESS_COUNT,
+    envvar=_env_name("PROCESS_COUNT"),
+    help="The amount of child processes to spawn.",
+    show_default=True,
+    type=int,
+)
+@click.option(
+    "--entrypoint-dir",
+    default=".",
+    envvar=_env_name("ENTRYPOINT_DIR"),
+    help="Look for the entrypoint's module in the specified directory.",
+    show_default=True,
+)
+def _run_cmd(  # pyright: ignore[reportUnusedFunction]
+    entrypoint: str,
+    token: str,
+    intents: hikari.Intents,
+    shard_count: int | None,
+    log_level: str,
+    process_count: int,
+    entrypoint_dir: str,
+) -> None:
+    sys.path.insert(0, entrypoint_dir)
+    logging.basicConfig(level=log_level.upper())
+    module_path, callback_name = entrypoint.split(":", 1)
+    callback = getattr(importlib.import_module(module_path), callback_name, "<NOT FOUND>")
+
+    if not callable(callback):
+        raise RuntimeError(f"{entrypoint!r} ({callback!r}) is not a function")
+
+    _service.run_subprocesses(
+        token, callback=callback, intents=intents, shard_count=shard_count, subprocess_count=process_count
+    )
+
+
+_SERVER_HELP = """
 Run a Hikari Orchestrator server instance.
+
 The `ADDRESS` for this server will default to TCP if no scheme is included and
 the valid schemes can be found at
 https://github.com/grpc/grpc/blob/master/doc/naming.md
 """
 
 
-@click.command(help=_HELP)
-@click.argument("address", default="localhost:0", envvar="ORCHESTRATOR_ADDRESS")
+@_cli_entry.command(name="server", help=_SERVER_HELP)
+@click.argument("address", default="localhost:0", envvar=_env_name("ADDRESS"))
 @click.option("--token", envvar="DISCORD_TOKEN", help="Discord token for the bot to orchestrate.", required=True)
 @click.option(
     "--intents",
     default=hikari.Intents.ALL_UNPRIVILEGED,
-    envvar="ORCHESTRATOR_INTENTS",
-    help="Gateway intents the bot should use. Defaults to ALL_UNPRIVILEGED",
+    envvar=_env_name("INTENTS"),
+    help="Gateway intents the bot should use.",
+    show_default=True,
     type=_cast_intents,
 )
-@click.option("--log-level", default="INFO", envvar="LOG_LEVEL", help="A Python logging level name. Defaults to INFO.")
+@click.option(
+    "--shard-count",
+    default=None,
+    envvar=_env_name("SHARD_COUNT"),
+    help="The amount of shards to run for the bot. Defaults to Discord's recommended amount.",
+    type=int,
+)
+@click.option("--log-level", default="INFO", envvar="LOG_LEVEL", help="A Python logging level name.", show_default=True)
 @click.option(
     "--ca-cert",
     default=None,
-    envvar="ORCHESTRATOR_CA_CERT",
-    help="Path to an unencrypted PEM certificate authority to use for encrypting TCP connections.",
+    envvar=_env_name("CA_CERT"),
+    help="System path to an unencrypted PEM certificate authority to use for encrypting TCP connections.",
     type=click.File("rb"),
 )
 @click.option(
     "--private-key",
     default=None,
-    envvar="ORCHESTRATOR_PRIVATE_KEY",
-    help="Path to an unencrypted PEM private key to use for authenticating TCP connections.",
+    envvar=_env_name("PRIVATE_KEY"),
+    help="System path to an unencrypted PEM private key to use for authenticating TCP connections.",
     type=click.File("rb"),
 )
-def _cli_entry(
+def _server_cmd(  # pyright: ignore[reportUnusedFunction]
     address: str,
     token: str,
-    ca_cert: io.BytesIO | None,
     intents: hikari.Intents,
+    shard_count: int | None,
     log_level: str,
+    ca_cert: io.BytesIO | None,
     private_key: io.BytesIO | None,
 ) -> None:
     logging.basicConfig(level=log_level.upper())
+
     if ca_cert:
         ca_cert_data = ca_cert.read()
         ca_cert.close()
@@ -116,7 +216,9 @@ def _cli_entry(
     else:
         private_key_data = None
 
-    _service.run_server(token, address, ca_cert=ca_cert_data, private_key=private_key_data, intents=intents)
+    _service.run_server(
+        token, address, ca_cert=ca_cert_data, private_key=private_key_data, intents=intents, shard_count=shard_count
+    )
 
 
 def main() -> None: